+++ src/lib/isNumeric.js
@@ -5,7 +5,7 @@ const numericNoSymbols = /^[0-9]+$/;
 
 export default function isNumeric(str, options) {
   assertString(str);
-  if (options && options.noSymbols) {
+  if (options && options.no_symbols) {
     return numericNoSymbols.test(str);
   }
   return numeric.test(str);